+++ src/swiss_ai_model_launch/launchers/model_catalog_entry.py
@@ -1,0 +1,15 @@
+from typing import Literal
+
+from pydantic import BaseModel
+
+
+class ModelCatalogEntry(BaseModel):
+    """A model entry from the catalogue — describes what the model needs, not how to run it."""
+
+    vendor: str
+    model_name: str
+    framework: Literal["sglang", "vllm"]
+    environment: str | None = None
+    nodes_per_worker: int = 1
+    framework_args: str | None = None
+    pre_launch_cmds: str | None = None
